CORE-4424 Port Link Manager to common worker framework
Delete legacy Link Manager application.
